Refactor ship rendering to new MUI Component Card
Started refactoring how the ships will be displayed into individual
cards.  Need to finish the typography for them and the styling of each
card.  I am also removing the info for each ship so that I can create a
module and route that will take the selected ships name and display all
the details on a seperate page.