Add @mui/icons-material and refactor sidebar
Sidebar refactored to be a top navbar for ease of access to links and
more aesthetic appeal.  mui icons added to dependencies for access to
their icons but might be removed later as I ended up not using the mui
nav bar. all non neccessary code from styled links removed.